fix(database): close client-quota leaks, sqlite cancel txn desync, postgres NUMERIC overflow
- pool: route PooledConnection::return_to_pool (sync + async) through Drop and
  have sync discard use safe_discard_connection, so the per-client quota slot is
  released. The explicit paths took the connection out of the guard and bypassed
  Drop's quota decrement, leaking the slot until the client wedged on
  ClientQuotaExceeded; return_to_pool also bypassed release_check and could
  re-pool poisoned connections. Removes the now-dead discard_connection helper.
- sqlite: a cancelled BEGIN / BEGIN IMMEDIATE / BEGIN EXCLUSIVE now marks
  NeedsRollback so the next operation's drain rolls back any transaction the
  sqlite3_interrupt race left open (a no-op when the connection is autocommit),
  preventing transaction-state desync / pool poisoning.
- postgres: widen NUMERIC digit-index arithmetic to i32 so an adversarial,
  unvalidated weight cannot overflow i16 (debug panic / release wrong digit) on
  hostile wire input.

Regression test: explicit_return_to_pool_and_discard_release_client_quota_slot.

+++ src/database/postgres.rs
@@ -7457,7 +7457,11 @@ fn decode_binary_numeric_to_text(data: &[u8]) -> Result<String, PgError> {
     }
 
     let digit_at_exponent = |exp: i16| -> u16 {
-        let idx = weight - exp;
+        // Widen to i32: `weight` is attacker-controlled and unvalidated, and in
+        // the fractional path `exp` is negative, so `weight - exp` can exceed
+        // i16::MAX (e.g. weight=0x7FFF) and overflow — a debug-build panic /
+        // release wrong-digit on hostile wire input.
+        let idx = i32::from(weight) - i32::from(exp);
         if idx < 0 {
             0
         } else {
